Data

Description

all datatypes for the algebraic trees and transformation the backstage stuff

Classes

class Eval e

the class of Eval denotes all things that are evaluatable to Rational - i used Rational as i want to be able to do algebraic transformations without rounding errors.

Types

type Label = Rational

Labels are chosen to be Rational so i can renumber and insert branches very cheaply - by subtracting half the denominator on the left branch and adding on the right branch respectively.

data Fun

The Fun(ction) datatype will represent - the function in a Node of the ExprTree expression tree

data Side

Denotes the side a branch will get inserted

Constructors

L

R

data Symbol

the Symbol datatype is - well for symbolic calculations

data Algebraic

the Alg datatype is a helper for reading the String input to make ExprTree

Constructors

Alg String

Functors

data ExprTree a

The standard Construction of a tree - with some extra flavour for expressions

Constructors

Leaf Label a
Leaf has a label and a value -

Node Label Fun (ExprTree a) (ExprTree a)
Node has a label used as above and a function - which should be used on the branches