Functions

Description

all internal Functions for manipulating ExprTrees also backstage fun

Functions

insert :: ExprTree a -> Label -> Fun -> Side -> ExprTree a -> ExprTree a

guess what - inserts a tree at label on given side into given tree

forkWith :: ExprTree a -> Fun -> Side -> ExprTree a -> ExprTree a

Just an insert on top level

relabel :: ExprTree a -> ExprTree a

relabels a tree on top level

relabelAt :: Label -> ExprTree a -> ExprTree a

relabels a tree startin on a given level

commutate :: ExprTree a -> ExprTree a

should be commutation laws for addition and multiplication

substitute :: [(Algebraic, Rational)] -> ExprTree Algebraic -> ExprTree Rational

substitutes a whole ExprTree according to a list of substitution rules

($$) :: (a1 -> a2, b1 -> b2) -> (a1, b1) -> (a2, b2)

applicates a pair of functions to a pair of values